Remove hidden connectivity methods usage
Network.java is a part of connectivity mainline module. The
hidden methods are no longer accessible outside the module.
The callers should either remove the usage or replace with
equivalent way.

The constructor of Network with a net id is hidden and not
accessible now for EpdgTunnelManagerTest. Replace the usage
with the mock network as other tests in EpdgTunnelManagerTest.

Bug: 182859030
Test: atest EpdgTunnelManagerTest#testTunnelSetupRequestParams
Change-Id: Ib40e39391fd57dd71f1a3004318ead505700d7c9