[iwlan] Fix incorrect datacallrsp
Since IKE does not allow to update network for session that is in
bringup, Iwlan service has an optimization to proactively "kill"
such sessions (and let framework try again).
There can be rare instances that IwlanDataService sends a request
for teardown to EpdgTunnelManager while it is still in the epdg
selection phase. In this case, no real "kill" or "close" api is
called, but TunnelManager still sends an onClose() CB to
DataService. This results in out of sync state between DataService
and TunnelManager...which results in subsequent call failures

Fix is to only call onClosed when the session is truly closed

Bug: 192569111
Test: wfc on device and atest
Change-Id: Ifc31f25ba823a29ce128f7d5f904caf74ebb12e3