Fix invalid test case #testWifiOnAvailable
This change will update this test case by verifying after calling
onCapabilitiesChanged() instead of onAvailable().
IwlanDataService sets the network as connected if
onCapabilitiesChanged() is invoked and NOP if onAvailable() is invoked.
But, this case is to verify network state change after onAvailable() is
called. The issue is hidden because other test cases set the network
state as connected before running this test case. And the current
IwlanDataService test won't re-create a clean object before runing each
test cases.

Bug: 235541193
Test: atest IwlanTests
Change-Id: Ia1e75cc20132d0792d1f95c84d39bdaba11f4fd8