Reset static instances before a test
Iwlan has static instances, and the instances might be initialized in
previous tests. The subsequent tests will unexpectedly fail because the
context is dirty. This change is to reset the static instance of SUT.
Then, the tests start in clean environments, and the results are
deterministic.

Bug: 228376556
Test: atest IwlanTests(new)
Change-Id: Ib9eac70fc9cdf0c1fbf4a2f8765d4c67973b22c6
Merged-In: Ib9eac70fc9cdf0c1fbf4a2f8765d4c67973b22c6